+++ src/frontend/src/lib/components/vip/VipQrCodeModal.svelte
