+++ src/frontend/src/lib/components/vip/VipRewardStateModal.svelte
