+++ src/frontend/src/tests/lib/components/address-book/AddressBookQrCodeScan.spec.ts
@@ -1,11 +1,11 @@
-import AddressBookQrCodeScan from '$lib/components/address-book/AddressBookQrCodeScan.svelte';
+import AddressBookQrCodeStep from '$lib/components/address-book/AddressBookQrCodeStep.svelte';
 import {
 	ADDRESS_BOOK_CANCEL_BUTTON,
 	ADDRESS_BOOK_QR_CODE_SCAN
 } from '$lib/constants/test-ids.constants';
 import { render } from '@testing-library/svelte';
 
-describe('AddressBookQrCodeScan', () => {
+describe('AddressBookQrCodeStep', () => {
 	const qrCodeContainerSelector = `div[data-tid="${ADDRESS_BOOK_QR_CODE_SCAN}"]`;
 	const cancelButtonSelector = `button[data-tid="${ADDRESS_BOOK_CANCEL_BUTTON}"]`;
 
@@ -22,7 +22,7 @@ describe('AddressBookQrCodeScan', () => {
 	});
 
 	it('should initialize and start scanning on mount', () => {
-		const { container } = render(AddressBookQrCodeScan, {
+		const { container } = render(AddressBookQrCodeStep, {
 			props: {
 				onClose: vi.fn(),
 				address: undefined
@@ -36,7 +36,7 @@ describe('AddressBookQrCodeScan', () => {
 
 	it('should trigger onClose when cancel button is clicked', () => {
 		const onClose = vi.fn();
-		const { container } = render(AddressBookQrCodeScan, {
+		const { container } = render(AddressBookQrCodeStep, {
 			props: {
 				onClose,
 				address: undefined
